Date functions other than tide utility move out from tide_utility
